+++ client/sdk/src/models/components/index.ts
@@ -174,6 +174,7 @@ export * from "./telemetrylogrecord.js";
 export * from "./tierlimits.js";
 export * from "./timeseriesbucket.js";
 export * from "./tool.js";
+export * from "./toolannotations.js";
 export * from "./toolcallsummary.js";
 export * from "./toolentry.js";
 export * from "./toolmetric.js";